requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=11.0.0
sqlalchemy>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
aiofiles>=23.0.0
